Move ContainerFlags to "commands" package
Commit 05513284e7f7559a9064acf25b302a28f5f040cf exposed the "rootfs"
and "no-pivot" flags for the "containers" command, but it accidentally
removed them for "run" since package-level variables are initialized
before package-level init functions in golang. Hoisting these flags to
a package imported by both commands solves the problem.
